test(data): move useCases to useCases folder
